Added accepting multiple execution args via more form fields
Fixes #39

This isn't as clean as I was hoping.

It works by bypassing the Rivets adapter on the ``arg`` iteration
binding then manually redrawing the list of args only when we're ready
(when the + button is pressed). Redrawing is handled by resyncing the
Rivets binding via a custom event.

This is needed because the default for-each-* binding destroys and
recreates all <input> elements which interrupts you as you're typing in
an input box. A more optimal solution would be to create a version of
for-each-* that reuses existing DOM elements. This will also be useful
down the road when we're dealing with very large collections.